working on Node.js support
