+++ default_plugins/links.js
@@ -1,0 +1,28 @@
+var knwl = require('../knwl');
+
+
+/* Link Parser */
+var Links = exports = module.exports;
+
+Links.calls = function() {
+    var results = [];
+    var words = knwl.words.linkWords;
+
+    for (var i = 0; i < words.length; i++) {
+        var word = words[i].replace(/[\(\)!]/g, ""); // replaces every bracket ')' or '(' and every '!' with an empty character
+        if (/^(https?|ftp):\/\/(-\.)?([^\s\/?\.#-]+\.?)+(\/[^\s]*)?$/i.test(word)) {
+            var link = word;
+            if (link[link.length - 1].search(/[?.!,]/g) !== -1) {
+                link = link.substr(0, link.length-1);
+            }
+            var linkObj = {
+                link: link,
+                preview: knwl.tasks.preview(i),
+                found: i
+            };
+
+            results.push(linkObj);
+        }
+    }
+    return results;
+};
