+++ package.json
@@ -8,7 +8,7 @@
     "parsing",
     "language"
   ],
-  "repository" : {
+  "repository": {
     "type": "git",
     "url": "https://github.com/loadfive/Knwl.js.git"
   },
@@ -21,10 +21,12 @@
     "node": "*"
   },
   "devDependencies": {
-    "jasmine-node": "*",
-    "uglify-js": "*"
+    "browserify": "^10.2.3",
+    "jasmine-node": "^1.14.5",
+    "uglifyjs": "^2.4.10"
   },
   "scripts": {
-    "build": "uglifyjs -o ./knwl.min.js ./knwl.js"
+    "build": "scripts/build.sh",
+    "test": "scripts/test.sh"
   }
 }
